refactor(system): align management pages with frontend standard
