+++ yak-ops-ui/src/pages/system/departments/types.ts
@@ -1,0 +1,7 @@
+export type DepartmentScope = 'all' | 'group' | 'leaf';
+
+export interface DepartmentTreeStats {
+  total: number;
+  groups: number;
+  leaves: number;
+}
