+++ yak-ops-ui/src/pages/system/oplogs/types.ts
@@ -1,0 +1,24 @@
+export interface OperationLogFilterValues {
+  operateType?: string;
+  operatePage?: string;
+  operationMethods?: string;
+  operator?: string;
+  operatorIp?: string;
+  target?: string;
+  targetType?: string;
+  detail?: string;
+  startTime?: number;
+  endTime?: number;
+}
+
+export type OperationLogSearchField =
+  | 'operator'
+  | 'target'
+  | 'detail'
+  | 'operatorIp';
+
+export interface OperationLogPaginationState {
+  current: number;
+  pageSize: number;
+  total: number;
+}
