+++ yak-ops-ui/src/pages/system/roles/types.ts
@@ -1,0 +1,23 @@
+export interface RolePaginationState {
+  current: number;
+  pageSize: number;
+  total: number;
+}
+
+export interface RoleFilterValues {
+  id?: number;
+  roleName?: string;
+  roleCode?: string;
+  description?: string;
+}
+
+export type RoleSearchField =
+  | 'roleName'
+  | 'roleCode'
+  | 'description'
+  | 'id';
+
+export interface RoleFormValues {
+  roleName: string;
+  description?: string;
+}
