+++ yak-ops-ui/src/pages/system/users/components/UserDetailDrawer.tsx
@@ -4,7 +4,6 @@ import {
   Space,
   Spin,
   Tag,
-  Typography,
   message,
 } from 'antd';
 import {
@@ -14,12 +13,13 @@ import {
   useState,
 } from 'react';
 
+import { YakEmpty } from '@/components/ui';
 import {
-  type SystemUser,
   getUserDetail,
+  type SystemUser,
 } from '@/services/security/users';
 
-import { errorText } from '../shared';
+import { getSystemErrorMessage } from '../../utils';
 
 export interface UserDetailDrawerRef {
   open: (user: SystemUser) => Promise<void>;
@@ -27,20 +27,22 @@ export interface UserDetailDrawerRef {
 
 const UserDetailDrawer = forwardRef<UserDetailDrawerRef>((_, ref) => {
   const [open, setOpen] = useState(false);
-  const [loading, setLoading] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
   const [detail, setDetail] = useState<SystemUser>();
 
   const show = useCallback(async (row: SystemUser) => {
     setOpen(true);
-    setLoading(true);
+    setIsLoading(true);
     setDetail(undefined);
 
     try {
       setDetail(await getUserDetail(row.id));
     } catch (error) {
-      
+      message.error(
+        getSystemErrorMessage(error, '用户详情加载失败'),
+      );
     } finally {
-      setLoading(false);
+      setIsLoading(false);
     }
   }, []);
 
@@ -59,65 +61,39 @@ const UserDetailDrawer = forwardRef<UserDetailDrawerRef>((_, ref) => {
       destroyOnClose
       onClose={close}
     >
-      <Spin spinning={loading}>
+      <Spin spinning={isLoading}>
         {detail ? (
           <Descriptions bordered column={1}>
-            <Descriptions.Item label="用户 ID">
-              {detail.id}
-            </Descriptions.Item>
-            <Descriptions.Item label="用户名">
-              {detail.userName}
-            </Descriptions.Item>
-            <Descriptions.Item label="真实姓名">
-              {detail.realName || '-'}
-            </Descriptions.Item>
-            <Descriptions.Item label="手机号">
-              {detail.phone || '-'}
-            </Descriptions.Item>
-            <Descriptions.Item label="邮箱">
-              {detail.email || '-'}
-            </Descriptions.Item>
+            <Descriptions.Item label="用户 ID">{detail.id}</Descriptions.Item>
+            <Descriptions.Item label="用户名">{detail.userName}</Descriptions.Item>
+            <Descriptions.Item label="真实姓名">{detail.realName || '-'}</Descriptions.Item>
+            <Descriptions.Item label="手机号">{detail.phone || '-'}</Descriptions.Item>
+            <Descriptions.Item label="邮箱">{detail.email || '-'}</Descriptions.Item>
             <Descriptions.Item label="角色">
               {detail.roleList?.length ? (
                 <Space wrap size={[4, 4]}>
                   {detail.roleList.map((role) => (
-                    <Tag color="blue" key={role.id}>
-                      {role.roleName}
-                    </Tag>
+                    <Tag key={role.id}>{role.roleName}</Tag>
                   ))}
                 </Space>
-              ) : (
-                '-'
-              )}
+              ) : '-'}
             </Descriptions.Item>
             <Descriptions.Item label="授权项目">
               {detail.projectList?.length ? (
                 <Space wrap size={[4, 4]}>
                   {detail.projectList.map((project) => (
                     <Tag key={project.id}>
-                      {project.projectName ||
-                        project.projectCode ||
-                        project.id}
+                      {project.projectName || project.projectCode || project.id}
                     </Tag>
                   ))}
                 </Space>
-              ) : (
-                '-'
-              )}
+              ) : '-'}
             </Descriptions.Item>
-            <Descriptions.Item label="创建时间">
-              {detail.createTime || '-'}
-            </Descriptions.Item>
-            <Descriptions.Item label="更新时间">
-              {detail.updateTime || '-'}
-            </Descriptions.Item>
+            <Descriptions.Item label="创建时间">{detail.createTime || '-'}</Descriptions.Item>
+            <Descriptions.Item label="更新时间">{detail.updateTime || '-'}</Descriptions.Item>
           </Descriptions>
         ) : (
-          !loading && (
-            <Typography.Text type="secondary">
-              暂无用户详情。
-            </Typography.Text>
-          )
+          !isLoading && <YakEmpty compact title="暂无用户详情" />
         )}
       </Spin>
     </Drawer>
@@ -125,5 +101,4 @@ const UserDetailDrawer = forwardRef<UserDetailDrawerRef>((_, ref) => {
 });
 
 UserDetailDrawer.displayName = 'UserDetailDrawer';
-
 export default UserDetailDrawer;
