+++ yak-ops-ui/src/pages/system/users/components/UserFilterBar.tsx
@@ -2,23 +2,29 @@ import {
   PlusOutlined,
   ReloadOutlined,
   SearchOutlined,
-} from "@ant-design/icons";
-import { Button, Input, Select, Tooltip } from "antd";
-import { useState } from "react";
+} from '@ant-design/icons';
+import { Input, Select, Tooltip } from 'antd';
+import { useMemo, useState } from 'react';
 
-import { PermissionGuard } from "@/components/security";
-import { SECURITY_PERMISSIONS } from "@/constants/securityPermissions";
+import { PermissionGuard } from '@/components/security';
+import {
+  YakButton,
+  YakFilterSwitch,
+  type YakFilterSwitchOption,
+} from '@/components/ui';
+import { SECURITY_PERMISSIONS } from '@/constants/securityPermissions';
 
-import type { RoleOption } from "../shared";
-
-export interface UserFilterValues {
-  id?: number;
-  userName?: string;
-  realName?: string;
-  roleId?: number;
-}
-
-type UserSearchField = "userName" | "realName" | "id";
+import {
+  ALL_ROLE_FILTER_VALUE,
+  USER_SEARCH_FIELD_OPTIONS,
+  USER_SEARCH_PLACEHOLDERS,
+} from '../constants';
+import type {
+  RoleOption,
+  UserFilterValues,
+  UserSearchField,
+} from '../types';
+import { buildUserFilters } from '../utils';
 
 interface UserFilterBarProps {
   roleOptions: RoleOption[];
@@ -27,35 +33,6 @@ interface UserFilterBarProps {
   onCreate: () => void;
 }
 
-const SEARCH_FIELD_OPTIONS: Array<{
-  label: string;
-  value: UserSearchField;
-}> = [
-  {
-    label: "用户名",
-    value: "userName",
-  },
-  {
-    label: "真实姓名",
-    value: "realName",
-  },
-  {
-    label: "用户 ID",
-    value: "id",
-  },
-];
-
-const SEARCH_PLACEHOLDERS: Record<UserSearchField, string> = {
-  userName: "请输入用户名",
-  realName: "请输入真实姓名",
-  id: "请输入用户 ID",
-};
-
-const clean = (value?: string): string | undefined => {
-  const normalized = value?.trim();
-  return normalized || undefined;
-};
-
 export default function UserFilterBar({
   roleOptions,
   onSearch,
@@ -63,131 +40,71 @@ export default function UserFilterBar({
   onCreate,
 }: UserFilterBarProps) {
   const [activeRoleId, setActiveRoleId] = useState<number>();
-  const [searchField, setSearchField] = useState<UserSearchField>("userName");
-  const [keyword, setKeyword] = useState("");
+  const [searchField, setSearchField] =
+    useState<UserSearchField>('userName');
+  const [keyword, setKeyword] = useState('');
 
-  const createFilters = (
-    nextKeyword = keyword,
-    nextSearchField = searchField,
-    nextRoleId = activeRoleId
-  ): UserFilterValues => {
-    const normalizedKeyword = clean(nextKeyword);
+  const roleFilterOptions = useMemo<
+    YakFilterSwitchOption<string>[]
+  >(
+    () => [
+      { value: ALL_ROLE_FILTER_VALUE, label: '全部' },
+      ...roleOptions.map((role) => ({
+        value: String(role.value),
+        label: role.label,
+      })),
+    ],
+    [roleOptions],
+  );
 
-    const filters: UserFilterValues = {
-      roleId: nextRoleId,
-    };
-
-    if (!normalizedKeyword) {
-      return filters;
-    }
-
-    if (nextSearchField === "id") {
-      const id = Number(normalizedKeyword);
-
-      if (Number.isInteger(id) && id > 0) {
-        filters.id = id;
-      }
-
-      return filters;
-    }
-
-    if (nextSearchField === "userName") {
-      filters.userName = normalizedKeyword;
-    }
-
-    if (nextSearchField === "realName") {
-      filters.realName = normalizedKeyword;
-    }
-
-    return filters;
-  };
+  const activeRoleValue =
+    activeRoleId === undefined
+      ? ALL_ROLE_FILTER_VALUE
+      : String(activeRoleId);
 
   const submit = () => {
-    onSearch(createFilters());
+    onSearch(buildUserFilters(keyword, searchField, activeRoleId));
   };
 
-  const changeRole = (roleId?: number) => {
-    setActiveRoleId(roleId);
+  const changeRole = (value: string) => {
+    const roleId =
+      value === ALL_ROLE_FILTER_VALUE ? undefined : Number(value);
+    const nextRoleId = Number.isFinite(roleId) ? roleId : undefined;
 
-    onSearch(createFilters(keyword, searchField, roleId));
+    setActiveRoleId(nextRoleId);
+    onSearch(buildUserFilters(keyword, searchField, nextRoleId));
   };
 
   const changeSearchField = (field: UserSearchField) => {
     setSearchField(field);
-    setKeyword("");
-
-    // 切换搜索字段时清除旧字段的查询条件，只保留角色筛选。
-    onSearch({
-      roleId: activeRoleId,
-    });
-  };
-
-  const changeKeyword = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const nextKeyword = event.target.value;
-
-    setKeyword(nextKeyword);
-
-    // 点击清空按钮后立即恢复到角色筛选状态。
-    if (!nextKeyword) {
-      onSearch({
-        roleId: activeRoleId,
-      });
-    }
+    setKeyword('');
+    onSearch(buildUserFilters('', field, activeRoleId));
   };
 
   return (
     <div className="mb-4 flex min-w-0 flex-col gap-4 lg:flex-row lg:items-center lg:justify-between">
-      {/* 左侧角色快捷筛选 */}
-      <div className="flex min-w-0 items-center gap-2 overflow-x-auto pb-1 lg:pb-0">
-        {[{ label: "全部", value: undefined }, ...roleOptions].map((role) => {
-          const active = activeRoleId === role.value;
-
-          return (
-            <div
-              key={role.value ?? "all"}
-              className={`
-          cursor-pointer
-          rounded
-          px-3
-          py-1
-          text-sm
-          font-medium
-          leading-5
-          transition-all
-          ${
-            active
-              ? "bg-[#f2f2f4] text-[#FE2C55]"
-              : `
-                bg-[#f2f4f7]
-                text-[#667085]
-                hover:bg-[#e8eaef]
-              `
-          }
-        `}
-              onClick={() => changeRole(role.value)}
-            >
-              {role.label}
-            </div>
-          );
-        })}
+      <div className="min-w-0 overflow-x-auto pb-1 lg:pb-0">
+        <YakFilterSwitch
+          value={activeRoleValue}
+          options={roleFilterOptions}
+          onChange={changeRole}
+        />
       </div>
 
-      {/* 右侧搜索和操作 */}
       <div className="flex shrink-0 flex-wrap items-center gap-2">
         <div className="flex h-8 w-[330px] max-w-full overflow-hidden rounded-md bg-[#f2f2f4]">
           <Select<UserSearchField>
             value={searchField}
-            options={SEARCH_FIELD_OPTIONS}
+            options={USER_SEARCH_FIELD_OPTIONS}
             variant="borderless"
             popupMatchSelectWidth={120}
             className={[
-              "h-8 w-[100px] shrink-0",
-              "[&_.ant-select-selector]:!h-8",
-              "[&_.ant-select-selector]:!px-3",
-              "[&_.ant-select-selector]:!bg-transparent",
-              "[&_.ant-select-selection-item]:!leading-[30px]",
-              "[&_.ant-select-selection-placeholder]:!leading-[30px]",
-            ].join(" ")}
+              'h-8 w-[100px] shrink-0',
+              '[&_.ant-select-selector]:!h-8',
+              '[&_.ant-select-selector]:!bg-transparent',
+              '[&_.ant-select-selector]:!px-3',
+              '[&_.ant-select-selection-item]:!leading-[30px]',
+            ].join(' ')}
             onChange={changeSearchField}
           />
 
@@ -197,26 +114,31 @@ export default function UserFilterBar({
             allowClear
             value={keyword}
             variant="borderless"
-            inputMode={searchField === "id" ? "numeric" : "text"}
-            placeholder={SEARCH_PLACEHOLDERS[searchField]}
+            inputMode={searchField === 'id' ? 'numeric' : 'text'}
+            placeholder={USER_SEARCH_PLACEHOLDERS[searchField]}
             suffix={
               <SearchOutlined
                 className="cursor-pointer text-slate-400 transition-colors hover:text-slate-700"
                 onClick={submit}
               />
             }
-            className={[
-              "min-w-0 flex-1",
-              "!h-8 !bg-transparent !py-0 !shadow-none",
-              "[&_.ant-input]:!bg-transparent",
-            ].join(" ")}
-            onChange={changeKeyword}
+            className="min-w-0 flex-1 !h-8 !bg-transparent !py-0 !shadow-none [&_.ant-input]:!bg-transparent"
+            onChange={(event) => {
+              const nextKeyword = event.target.value;
+              setKeyword(nextKeyword);
+              if (!nextKeyword) {
+                onSearch(
+                  buildUserFilters('', searchField, activeRoleId),
+                );
+              }
+            }}
             onPressEnter={submit}
           />
         </div>
 
         <Tooltip title="刷新列表">
-          <Button
+          <YakButton
+            iconOnly
             icon={<ReloadOutlined />}
             onClick={onRefresh}
           />
@@ -226,13 +148,13 @@ export default function UserFilterBar({
           mode="one"
           permission={SECURITY_PERMISSIONS.user.create}
         >
-          <Button
+          <YakButton
             type="primary"
             icon={<PlusOutlined />}
             onClick={onCreate}
           >
             新增用户
-          </Button>
+          </YakButton>
         </PermissionGuard>
       </div>
     </div>
